move algorithm oflder
