+++ packages/algorithm/package.json
@@ -1,0 +1,46 @@
+{
+  "name": "@geoarrow/algorithm",
+  "version": "0.4.0-beta.0",
+  "description": "",
+  "type": "module",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    }
+  },
+  "sideEffects": false,
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsc --build tsconfig.build.json",
+    "test": "vitest run",
+    "test:watch": "vitest",
+    "typecheck": "tsc --noEmit",
+    "prepublishOnly": "npm run build"
+  },
+  "keywords": [],
+  "author": "Development Seed",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/geoarrow/geoarrow-js.git"
+  },
+  "devDependencies": {
+    "@types/node": "^25.3.3",
+    "jsdom": "^28.1.0",
+    "typescript": "^6.0.2",
+    "vitest": "^4.0.18"
+  },
+  "dependencies": {
+    "apache-arrow": ">=15",
+    "@geoarrow/schema": "workspace:^"
+  },
+  "peerDependencies": {},
+  "volta": {
+    "extends": "../../package.json"
+  }
+}
