+++ packages/algorithm/src/earcut.ts
