+++ packages/algorithm/src/exterior.ts
