+++ packages/algorithm/src/index.ts
