+++ packages/algorithm/src/proj.ts
@@ -1,6 +1,6 @@
 import * as arrow from "apache-arrow";
 import proj4 from "proj4";
-import type { GeoArrowType } from "../type";
+import type { GeoArrowType } from "../../schema/src/type";
 import { mapCoords } from "./coords";
 
 /**
