+++ packages/algorithm/src/utils/assert.ts
