+++ packages/algorithm/src/utils/polygon.ts
