+++ packages/algorithm/src/winding.ts
