--- src/index.ts
@@ -1,7 +1,0 @@
-export * as type from "../packages/schema/src/type.js";
-export * as algorithm from "./algorithm";
-export * as child from "./child.js";
-export * as data from "./data.js";
-export * as geoInterface from "./geo-interface";
-export * as vector from "./vector.js";
-export * as worker from "./worker";
